v0.0.44-alpha: 修复多语言支持和bilibili duration类型问题

+++ backend/app/models/schemas.py
@@ -55,6 +55,6 @@ class VideoFormat(BaseModel):
 class VideoInfo(BaseModel):
     title: str
     thumbnail: Optional[str] = None
-    duration: Optional[int] = None
+    duration: Optional[float] = None
     uploader: Optional[str] = None
     formats: list[VideoFormat] = []
